changed inq from int casting

+++ NuSysApp/Views/Misc/InqCanvas/InqLine.xaml.cs
@@ -166,7 +166,7 @@ namespace NuSysApp
                                 {
                                     string[] coords = p.Split(new string[] { "," }, StringSplitOptions.RemoveEmptyEntries);
                                     //Point point = new Point(double.Parse(coords[0]), double.Parse(coords[1]));
-                                    Point parsedPoint = new Point(Int32.Parse(coords[0]), Int32.Parse(coords[1]));
+                                    Point parsedPoint = new Point(double.Parse(coords[0]), double.Parse(coords[1]));
                                     line.AddPoint(parsedPoint);
                                 }
                             }
